05_model/requirements.txt
# Core
pandas>=2.1
numpy>=1.26
scikit-learn>=1.3
scipy>=1.11
tqdm>=4.66
typer>=0.9
hydra-core>=1.3
pyyaml>=6.0

# Bio / ontology
biopython>=1.81
obonet>=0.6
networkx>=3.2

# Torch (CPU by default; change if you install CUDA wheels)
torch>=2.1
torchvision>=0.16
torchaudio>=2.1

# Viz
matplotlib>=3.8
seaborn>=0.13

# Kaggle API
kaggle>=1.6
